Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -2,7 +2,7 @@
 int main() {
     char a;
     scanf("%c",&a);
-    printf("ASCII Value: %c",a);
+    printf("ASCII Value: %c",(int)a);
 
     return 0;
 }
